% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/euclideanmetric.R
\name{euclideanmetric}
\alias{euclideanmetric}
\title{Calculate the euclideanmetric distance of two points in the real number space}
\usage{
euclideanmetric(a, b)
}
\arguments{
\item{a}{A vector, the coordinate of point a in real number space, in this package, it represents the delta of two stable isotope}

\item{b}{Just the paramater a}
}
\description{
This function is the core function of this package, which could calculate the 'distance' between sources and target mixture
}
\examples{
rho <- euclideanmetric(a = c(3,4), b = c(1, 2))
cat("The euclideanmetric distance is")
print(rho)
}

