% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iso.euclidean2.R
\name{iso.euclidean2}
\alias{iso.euclidean2}
\title{Generate the sample list of your input}
\usage{
iso.euclidean2(source.matrix = 0, mixture.matrix = 0,
  correctiso.matrix = NULL)
}
\arguments{
\item{source.matrix}{the data of source, must be a data.frame}

\item{mixture.matrix}{the data of mixture, must be a data.frame}

\item{correctiso}{the data of correctiso matrix, must be a data.frame}
}
\description{
Running this function is always the first step of this model, a sample list, which contains the euclideanmetric distance between the mixture and sources is the output of this function
iso.euclidean2(source.matrix = 0, mixture.matrix = 0, correctiso.matrix = NULL)
}
\examples{
source <- data.frame(d15NPl = c(6.50, 4.42, 11.19, 9.82), d13CPl = c(-11.17, -30.88, -10.19, -15.01), sd.N = c(1.4594632, 2.2680709, 1.1124385, 0.8271039), mean.N = c(6.488984, 4.432160, 11.192613, 9.816280), sd.C = c(1.2149562, 0.6413182, 1.9593306, 1.1724677), mean.C = c(-11.17023, -30.87984, -11.17090, -14.05701))
mixture <- data.frame(d15NPl = 10.30, d13CPl = -11.58)
samplelist <- iso.euclidean2(source, mixture)
print(samplelist[[1]][1:10, ])
}

