% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iso.histograms.R
\name{iso.histograms}
\alias{iso.histograms}
\title{Draw a histograms using the output of function iso.mcm()}
\usage{
iso.histograms(mcm.result, numbers = 0)
}
\arguments{
\item{mcm.result}{the result of function}

\item{numbers}{the numbers of your sources}
}
\description{
This funcition draw a histograms if your input the correct paramaters
iso.histograms(mcm.result, numbers = 0)
}
\examples{
source.matrix <- data.frame(d15NPl = c(6.50, 4.42, 11.19, 9.82), d13CPl = c(-11.17, -30.88, -10.19, -15.01), sd.N = c(1.4594632, 2.2680709, 1.1124385, 0.8271039), mean.N = c(6.488984, 4.432160, 11.192613, 9.816280), sd.C = c(1.2149562, 0.6413182, 1.9593306, 1.1724677), mean.C = c(-11.17023, -30.87984, -11.17090, -14.05701))
mixture.matrix <- data.frame(d15NPl = 10.30, d13CPl = -11.58)
sample.list <- iso.euclidean2(source.matrix, mixture.matrix)
mcm.result <- iso.mcm(sample.list, numbers = 4)
iso.histograms(mcm.result, 4)
}

