% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iso.loaddata.R
\name{iso.loaddata}
\alias{iso.loaddata}
\title{Load the data from R objects or files}
\usage{
iso.loaddata()
}
\description{
This function can guide the user load data from R objects or files, don't forget to use '<-' to assign a object
}
\examples{
datalist <- iso.loaddata()
}

