% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iso.mcm.R
\name{iso.mcm}
\alias{iso.mcm}
\title{Resample the sample list using random sampling method}
\usage{
iso.mcm(source.list, numbers, run = 1000)
}
\arguments{
\item{source.list}{the output of function iso.euclidean2}

\item{numbers}{the numbers of your sources}

\item{run}{how many resample cycle do you want? actually run = 1000 is enough for most situation}
}
\description{
This function is kind of resample function
iso.mcm(source.list, numbers, run = 1000)
}
\examples{
source.matrix <- data.frame(d15NPl = c(6.50, 4.42, 11.19, 9.82), d13CPl = c(-11.17, -30.88, -10.19, -15.01), sd.N = c(1.4594632, 2.2680709, 1.1124385, 0.8271039), mean.N = c(6.488984, 4.432160, 11.192613, 9.816280), sd.C = c(1.2149562, 0.6413182, 1.9593306, 1.1724677), mean.C = c(-11.17023, -30.87984, -11.17090, -14.05701))
mixture.matrix <- data.frame(d15NPl = 10.30, d13CPl = -11.58)
sample.list <- iso.euclidean2(source.matrix, mixture.matrix)
mcm.table <- iso.mcm(sample.list.list, 4)
print(mcm.table[1:10, ])
}

