% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iso.plot.R
\name{iso.plot}
\alias{iso.plot}
\title{Draw the dot plot of mixture and sources}
\usage{
iso.plot(datalist = list(), group = 0)
}
\arguments{
\item{datalist}{the output of iso.loaddata}
}
\description{
This funcion is used to draw a dot plot of your source and mixture, use with function iso.loaddata usually
iso.plot(datalist = list(), group = 0)
}
\examples{
source.matrix <- data.frame(d15NPl = c(6.50, 4.42, 11.19, 9.82), d13CPl = c(-11.17, -30.88, -10.19, -15.01), sd.N = c(1.4594632, 2.2680709, 1.1124385, 0.8271039), mean.N = c(6.488984, 4.432160, 11.192613, 9.816280), sd.C = c(1.2149562, 0.6413182, 1.9593306, 1.1724677), mean.C = c(-11.17023, -30.87984, -11.17090, -14.05701))
mixture.matrix <- data.frame(d15NPl = 10.30, d13CPl = -11.58)
iso.plot(datalist = list(mixture = mixture.matrix, sources = source.matrix))
}

